refactor: rename test file extension to .test.js
